Video publishing consolidated into one hook.
